src/database/cloudSqlService.ts
import { createLog, fetchLogs, fetchLogsByEventType, fetchLogsByUserId, fetchLogsByTimestamp, fetchLogsByTimestampRange, deleteLogsByEventType, deleteLogsByTimestamp, deleteLogsByTimestampRange } from './cloudSqlService.js';

// ... existing code ...

export const fetchLogsByEventType = async (eventType: string) => {
  // Implement logic to fetch logs by event type from Cloud SQL
  console.log('Fetching logs by event type:', eventType);
  return [
    // Example log data
    { id: 'log1', userId: 'user1', eventType: 'login', timestamp: '2023-10-26T10:00:00Z', data: 'Successful login' },
    { id: 'log2', userId: 'user2', eventType: 'report_created', timestamp: '2023-10-27T12:30:00Z', data: 'New report created' },
  ];
};
